+++ HISTORY.rst
@@ -22,6 +22,7 @@ latest (unreleased)
 **Features and Improvements**
 
 * BSSFL-516: Team on journal items list view
+* BSSFL-514: Budget with project code
 * BSSFL-488: Change RMA rules for Picking generation. The conditions for
   generating the procurement order have been changed if he is linking to a rma.
 * BSSFL-499: Generate picking (on conditions) when down payment is paid